requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
#plotly==5.17.0
dash
folium==0.14.0
#numpy==1.24.3
#pandas==2.0.3
gunicorn==21.2.0
flask-wtf==1.1.1
sqlalchemy==2.0.23
mysqlclient==2.2.7
sqlalchemy_utils==0.41.1
statistics==1.0.3.5
cx_Oracle==8.3.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Visualizações avançadas
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dashboard interativo
streamlit>=1.25.0
streamlit-autorefresh>=0.0.1

# APIs climáticas
requests>=2.31.0
python-dateutil>=2.8.0

# Processamento de dados
scipy>=1.11.0
joblib>=1.3.0

# Otimizações
numba>=0.57.0  # JIT compilation para performance
cachetools>=5.3.0  # Cache avançado

# Monitoramento
psutil>=5.9.0  # Monitoramento de sistema
py-cpuinfo>=9.0.0  # Informações de hardware